feat(signals): T3 — stacked multi-widget layout with synced time axis
Converts the single-chart Signals page into a vertical stack of
independent chart widgets. New SignalWidget owns its own query AST,
chart type, fetch, loading/error/latency state — so one widget behaves
exactly like the old page. SignalsPage now keeps only shared concerns
(vehicle, timeframe + picker, the collected-signals table) plus a widget
registry with add / delete / hide. All panels join one ECharts
connection group ("signals-page") via a new additive `groupId` prop on
QueryChart, so the hover cursor + tooltip sync across panels; the
no-groupId path is unchanged. Brush-select on any panel bubbles up to set
the shared timeframe and every widget refetches.
